+++ torchaudio/datasets/__init__.py
@@ -2,6 +2,6 @@ from .commonvoice import COMMONVOICE
 from .librispeech import LIBRISPEECH
 from .vctk import VCTK
 from .yesno import YESNO
-from .utils import DiskCache
+from .cache import DiskCache
 
 __all__ = ("COMMONVOICE", "LIBRISPEECH", "VCTK", "YESNO", "DiskCache")
